detect availablity of 'long long'
cast time_t to 'long long' prior to bitshifting since it can be a float value according to POSIX.1-2008

closes #3439

+++ init.c
@@ -2868,8 +2868,11 @@ static void mutt_srandom (void)
   /* POSIX.1-2008 states that seed is 'unsigned' without specifying its width.
    * Use as many of the lower order bits from the current time of day as the seed.
    * If the upper bound is truncated, that is fine.
+   *
+   * tv_sec is integral of type integer or float.  Cast to 'long long' before
+   * bitshift in case it is a float.
    */
-  seed = (tv.tv_sec << 20) | tv.tv_usec;
+  seed = ((LONGLONG) tv.tv_sec << 20) | tv.tv_usec;
   srandom(seed);
 }
